requirements.txt
Flask
flask-cors
Pillow
numpy
gunicorn
# Ini library ajaib agar ringan di Render (Linux Only)
https://github.com/google-coral/pycoral/releases/download/v2.0.0/tflite_runtime-2.5.0.post1-cp39-cp39-linux_x86_64.whl